services/python/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
asyncpg==0.29.0
alembic==1.13.2
pydantic==2.9.2
gpxpy==1.6.2
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator
bcrypt==3.2.2
